Support for unreleased WIP
- TODO: move back test_utils with clast
- TODO: guaranteed unreleased parsed only once...
Review the architecture. The line between "parsing" and "validation" is starting to blur.
Let's dumb down the parsing to something as simple as just mapping dmast nodes to clast nodes.
We can resolve those and validate all of it in the next steps. Let's make the crates even smaller.
Something like: 1-to-1 mapping of mdast, sequencing (formatting?), linting
- TODO: Fix main crate compilation
- TODO: Use spans at node level, and location in errors?
- TODO: Debug, Clone, PartialEq, Eq errwhere
- TODO: The to_string() method or, something similar, should use the original string combined with the
location.
- TODO: canonical form is a bit different. This one could reformat.